refactor(client): move host-connection UI out of kaval/ into host/ (pure motion)
