+++ packages/client/src/App.tsx
@@ -48,9 +48,9 @@ import { createImportSessionAction } from "./importSessionAction";
 import { useShortcuts } from "./input/useShortcuts";
 import IntentEditorDialog from "./intent/IntentEditorDialog";
 import { useIntentEditor } from "./intent/useIntentEditor";
-import { ConnectCanvas } from "./kaval/ConnectCanvas";
+import { ConnectCanvas } from "./host/ConnectCanvas";
 import DegradedCanvas from "./kaval/DegradedCanvas";
-import HostDownCanvas from "./kaval/HostDownCanvas";
+import HostDownCanvas from "./host/HostDownCanvas";
 import { type CanvasMode, canvasMode } from "./kaval/useCanvasMode";
 import MobileKeyBar from "./MobileKeyBar";
 import MobileTileView from "./MobileTileView";
