+++ packages/client/src/host/connectCanvasCopy.test.ts
