+++ packages/client/src/host/connectCanvasCopy.ts
